src/backend/requirements.txt
# Web 框架
fastapi~=0.115
uvicorn[standard]~=0.34
python-socketio~=5.11
python-multipart~=0.0.18

# AI/ML
transformers>=4.45,<4.50
torch>=2.1,<2.7
# 注意: torch 需要通过 --index-url 安装 CUDA 版本
# pip install torch --index-url https://download.pytorch.org/whl/cu121
torchvision>=0.16,<0.22
accelerate~=1.2
# vllm>=0.3  # 仅 Linux，Windows 自动降级到 transformers

# 向量数据库
chromadb~=0.5

# 音频
sounddevice~=0.5
soundfile~=0.12
numpy>=1.24,<2.0

# HTTP
httpx~=0.28

# 图像
Pillow~=11.1
mss~=9.0

# 定时任务
apscheduler~=3.10

# 配置/工具
pyyaml~=6.0
pydantic~=2.10
psutil~=6.1
gputil
